// frontend-template/src/api/modules/enums/menu-type-enum.ts
/**
 * Fast.Admin.Core.Enum.Menus.MenuTypeEnum 菜单类型枚举
 * @export
 * @enum {string}
 */
export enum MenuTypeEnum {
    /**
     * 目录
     */
    Catalog = 1,
    /**
     * 菜单
     */
    Menu = 2,
    /**
     * 内链
     */
    Internal = 3,
    /**
     * 外链
     */
    Outside = 4,
}